Id de Granja cambiado a Long para que funcione Delete

+++ src/main/java/com/accenture/granja/repository/GranjaRepository.java
@@ -3,7 +3,7 @@ package com.accenture.granja.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.accenture.granja.model.Granja;
-public interface GranjaRepository extends JpaRepository<Granja, Integer>{
+public interface GranjaRepository extends JpaRepository<Granja, Long>{
 
 
 }
